datId moved to relation info
